Moved graph controls into its own module.
